Move sambusac test to different package instead of e2e
