+++ Dockerfile.build
@@ -12,4 +12,4 @@ RUN CGO_ENABLED=0 go build -o orbs-node -a -ldflags '-extldflags "-static"' main
 
 RUN CGO_ENABLED=0 go build -o orbs-json-client -a -ldflags '-extldflags "-static"' devtools/jsonapi/main/main.go
 
-RUN CGO_ENABLED=0 go build -o sambusac -a -ldflags '-extldflags "-static"' devtools/sambusac/main.go
+RUN CGO_ENABLED=0 go build -o sambusac -a -ldflags '-extldflags "-static"' devtools/sambusac/main/main.go
